packages/graphrecon-lib/pyproject.toml
[project]
name = "graphrecon-lib"
version = "0.1.0"
description = "Shared library for all graphrecon tools"
readme = "README.md"
authors = [
    { name = "Julien OPPLIGER", email = "account@oppliger.cc" }
]
requires-python = ">=3.13"
dependencies = [
    "open-asset-store>=0.0.2",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"
